Refactor data shared by rss pages
