+++ apps/website/src/app/feeds/book-club.xml/route.ts
@@ -1,75 +1,8 @@
 import { Feed } from "feed";
 
-import type { PartialDateString } from "@alveusgg/data/build/types";
-
 import { env } from "@/env";
 
-// import { typeSafeObjectEntries } from "@/utils/helpers";
-// import { camelToKebab } from "@/utils/string-case";
-
-// TODO move BookInfo & books to shared location
-
-type BookInfo = {
-  title: string;
-  author: string;
-  // image: StaticImageData;
-  month: PartialDateString & `${number}-${number}`;
-  link: string;
-  // thickness: (typeof thickness)[keyof typeof thickness];
-  // color: `border-${string}`;
-  vodId?: string;
-};
-
-const books: BookInfo[] = [
-  {
-    title: "The Anthropocene Reviewed",
-    author: "John Green",
-    // image: theAnthropoceneReviewed,
-    month: "2025-05",
-    link: "https://amzn.to/42LpJSo",
-    // thickness: thickness.sm, // 300 pages
-    // color: "border-black",
-  },
-  {
-    title: "The Last Rhinos",
-    author: "Lawrence Anthony",
-    // image: theLastRhinos,
-    month: "2025-04",
-    link: "https://amzn.to/3YibG59",
-    // thickness: thickness.md, // 360 pages
-    // color: "border-alveus-tan-900",
-  },
-  {
-    title: "Adventures of a Young Naturalist",
-    author: "Sir David Attenborough",
-    // image: adventuresOfAYoungNaturalist,
-    month: "2025-03",
-    link: "https://amzn.to/41qA2um",
-    // thickness: thickness.lg, // 400 pages
-    // color: "border-blue-900",
-    vodId: "fRjxSfOhk94",
-  },
-  {
-    title: "A Most Remarkable Creature",
-    author: "Jonathan Meiburg",
-    // image: aMostRemarkableCreature,
-    month: "2025-02",
-    link: "https://amzn.to/410hD8x",
-    // thickness: thickness.lg, // 400 pages
-    // color: "border-alveus-tan-200/75",
-    vodId: "AH0mwSckVns",
-  },
-  {
-    title: "H is for Hawk",
-    author: "Helen Macdonald",
-    // image: hIsForHawk,
-    month: "2025-01",
-    link: "https://amzn.to/4a2ByGQ",
-    // thickness: thickness.sm, // 320 pages
-    // color: "border-black",
-    vodId: "EfCksCFlq84",
-  },
-];
+import books from "@/data/books";
 
 export async function GET() {
   const bookClubPageUrl = `${env.NEXT_PUBLIC_BASE_URL}/book-club`;
